# workbench/docs/lab-openapi.yaml
openapi: 3.1.0
info:
  title: Svoe Vino lab API
  version: "1.0.0"
  summary: The HTTP contract of pipeline/lab_server.py.
  description: |
    This document describes the local Svoe Vino laboratory on port 8168.
    The server reads and writes the laboratory catalogue, embedding artifacts, test
    sets, and run artifacts. It has no authentication. Keep it on a trusted local host.

    JSON write operations use UTF-8. Image upload operations send raw image bytes.
    `HEAD` has the same status and headers as `GET` for every documented GET route.
  license:
    name: Not stated
    identifier: LicenseRef-not-stated
servers:
  - url: http://127.0.0.1:8168
    description: Local lab server. The port comes from `--port`; the default is 8168.
...
paths:
...
  /api/embeddings:
    get:
      tags: [embeddings]
      operationId: listEmbeddings
      summary: Read every embedding configuration, item counts, and build state.
      responses:
        "200": { $ref: "#/components/responses/Object" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/embeddings/{name}:
    get:
      tags: [embeddings]
      operationId: getEmbedding
      summary: Read all wines and item cells of one embedding.
      parameters:
        - $ref: "#/components/parameters/Name"
      responses:
        "200": { $ref: "#/components/responses/Object" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  parameters:
...
    Name:
      name: name
      in: path
      required: true
      schema: { type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]*$" }
...
  responses:
    Object:
      description: JSON object.
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Error:
      description: Error response.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string }
...